requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
google-cloud-aiplatform
pyyaml